Cleanup workspace
The recommendation for a Cargo workspace is that the top level package
should be empty except for a list of packages within the workspace.
This enables features such as multiple binaries within the workspace,
which is something we will require soon.
It should also generally reduce the tedium of maintenance.
