+++ av1an-core/Cargo.toml
@@ -87,6 +87,7 @@ quickcheck = { version = "1.0.3", default-features = false }
 quickcheck_macros = "1"
 
 [features]
+default = ["vapoursynth_new_api"]
 ffmpeg_static = ["ffmpeg/static", "ffmpeg/build"]
 vapoursynth_new_api = [
   "vapoursynth/vapoursynth-api-32",
